cf_code/requirements.txt
google-cloud==0.34.0
google-cloud-bigquery==2.33.0